+++ CHANGELOG.rst
@@ -6,6 +6,15 @@ All notable changes to the MEF_ELine NApp will be documented in this file.
 [Unreleased]
 ************
 
+[2023.2.10] - 2025-02-12
+************************
+
+Fixed
+=====
+- Parametrized requests timeout when finding paths and when sending flow mods to avoid potentially hanging indefinitely
+- Fixed Path ``choose_vlans`` to be all or nothing, if a path link fails to allocate a vlan, it'll release the allocated vlans.
+
+
 [2023.2.9] - 2025-01-22
 ***********************